Begin sketching out image operations
Image operations refer to the transformations a user can apply to a
screenshot they are taking. GUI support is left for later.

The plan is to have a stack of them and apply them to an image from the
oldest operation to the newest one on draw events.

This patch adds a bunch of structs useful in drawing (Point, Rectangle,
Colour...) and implements the `DrawEllipse` operation as I believed it
would be a difficult one to do. I receievd a bunch of help with the
ellipse from tarche and Arkhist.

Co-authored-by: Arkhist <peyrille.benjamin@gmail.com>