Make it possible to edit a screenshot
This patch adds a toolbar containing various buttons for the various
actions one might find useful when editing a screenshot. Currently
these buttons have text indicating what their purpose is, as I don't
have icons for them.

This works by hooking on various events on the DrawingArea we use to
show the screenshot (notify-motion-event, button-release, button-press)
to communicate to the OperationStack information about the Operation on
top being modified in various ways. OperationStack is heavily
restructured as a consequence (it was likely pushed far too soon).

Furthermore a concept of `Tool`s is added, which are dataless
operations is added, such that it is easier to communicate between the
buttons and the OperationStack.