Add colour chooser buttons to the toolbar
This patch allows a user to select colours they want for border, and
fill of a given shape.

We now track a primary and secondary colour inside the OperationStack
that gets used when creating new operations. As such
`Operation::create_default_for_tool` is extended to take the primary and
secondary colours as parameters.

These buttons receive special "image" widgets that are actually
`DrawingArea` which draw suggestive imagery wrt which colour the button
modifies. The fill button draws a checkerboard when the selected colour
is fully transparent.