Make the history model implementation (hopefully) correct
This rather intrusive changes makes the model report the number of items
in the sqlite db as its `n_items`, we still fetch new items in
`item` but that should be fine I believe with how we do things now.

We also now emit a `items-changed` signal when an item is inserted in
the history.

This should fix #11 and remove a few items from my yakstack :^)