Add window select mode to the crop tool :^)
This patch adds the ability for the user to select a window when
cropping. This is done by retrieving the window rects from the x
server and then using that as crop region when clicking on the editor.
Furthermore the user may choose between selecting windows including
decorations, windows excluding decorations (i.e. only the content of
the window) and not selecting windows at all.

I'm honestly not sure the logic is implemented in the best way, but it
works OK for now.

Thanks Andreas Björkman(https://github.com/CocaineJohnsson) for
helping me figure out how other tools retrieve window coordinates and
being generally helpful.
Thanks tarche for being a good rubber duck. :)