Fix crash after taking a screenshot and navigating the main window
It appears that trying to call `HistoryModel::items_changed` inside a
PostCaptureAction was the cause of  #11. This happened indirectly
through the now removed `HistoryModel::add_item_to_history`. This has
been replaced with a mechanism that attempts to call `items_changed`
from the mainloop by using channels. This seems to work in preliminary
testing, so I'm saying it's good.

This change is a bit intrusive, but it also makes `EditorWindow` no
longer store a `HistoryModel`, and ensures only one exists per `KCShot`
instance by creating it inside `KCShot::constructed`.

We also now close the EditorWindow before saving to remove a bit of
lagginess in debug mode.

Fixes #11.