Fix rectangle tools running away from the cursor in specific situations
This would happen when the user would start a rectangle tool with the
bottom-right corner and then went up or to the left. This would
eventually cause a crash when trying to save the screenshot or save
garbage.

This patch also replaces `Rectangle::{normalise => normalised}` as the
former was awkward to use.