Ensure that a fullscreen window's content_rect and outer_rect are equal
This patch moves out content_rect retrieval into its own function and
then ensures that the outer_rect will be the same as the content_rect
for fullscreen windows. This is necessary as some window managers
(for example awesomewm) return frame extents even for fullscreen windows
and that could lead to garbage being captured.