Be more mindful about object lifetimes in colour button code
"Object lifetimes" refers to gobject ref-counting in the title, not Rust
lifetimes.

This patch makes the callbacks not take any strong references any more.
From now on we use weak references and/or moves.

This is more work on the memory leak issue, I think we're on a good
track because looking in htop, less memory seems to be leaked on
average.