Add "crop-first" mode to kcshot
This allows user to configure kcshot so that editing a screenshot begins
with selecting the area they wish to edit. A current limitation is that
you cannot resize the cropped area, however it will be lifted in the
future.

This mode will further be useful in the future when we may want to add
QR code scanning or OCR-based features.