Ensure the "present-hide main window" hacky dance happens at most once
In 5367569839654b77e7f6574c1b100879c0edf86d I made it so we always
presented/hid the main window to ensure we could obtain an ashpd window
identifier (not doing so results in a crash). This is a bit inefficient
as we only need the window to be shown _once_ for WindowIdentifier
creation to run succesfully. This commit fixes that.