CI: Remove manual handling of `skip ci` surrounded in `[]`
I added this in the workflow of a previous project so I could skip CI
when I wasn't yet done with something, but I wanted to push a PR anyway,
and didn't want to waste CI resources. It seems Github actions can do
this natively [1] now, so there's no point in having this anymore.

[1]: https://github.blog/changelog/2021-02-08-github-actions-skip-pull-request-and-push-workflows-with-skip-ci/