CI: Only attempt to install `cargo-cranky` if it doesn't exist
`cargo install` refuses to install a package if it already exists
(which is good -- let's not do more work than necessary in CI), but it
does so by returning 1, which causes the whole workflow to fail. This
should stop `cargo install` from running at all if `cargo cranky` is
already installed.