Refactor `EditorWindow` to make it less spaghetti
Previously we did a lot of the work in `constructed` which made that
function quite long and hard to read. This patch splits it in logical
chunks to make the code as a whole more readable.

This patch also moves a lot of stuff to be handled by blueprints, which
makes code nicer.