Update dependencies
This ended up needing a `cargo update` because otherwise I got errors in
ashpd and zbus? Not sure what that was about.

Also changed xcb to use crates.io instead of git while we were at it as
the reason why we used git in the first place got released.

Only code changes in this commit are ones necessary for code to compile,
no QoL improvements.