Use target surface in pixelate and blur operations
This ensures that during editing-mode, the user will see the correct
blurred/pixelated image rather than the original image with those
operations applied. This means that now e.g. text will be correctly
blurred.

This patch also updates pixelops::{blur, pixelate} to handle alpha, as
it was required to implement this.

Fixes: #3